requirements.txt
Django==1.8
argparse==1.2.1
decorator==4.0.6
django-extensions==1.6.1
djangorestframework==3.3.2
drfdocs==0.0.6
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
jsonfield==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wsgiref==0.1.2
